Create mathKing.py
The game “Math King”, gives the user the possibility to start a new game in two ways, the only button available in the “New” button, so by clicking on “New” or by pressing Enter on the keyboard the user can start a new game. 
Then the user can type the solution for num1 + num2 or num1 – num2 and the only button available to press is the “Submit” button or instead press Enter on the keyboard to achieve the same result.
When right, labels turn green and the user can hear a nice pling; instead, when the user is wrong, labels turn yellow and red, and the user can hear a downgoing pling. Different messages are shown depending on the results. Score updates each time and the best score is stored.
“New” button clean the answer box.
The info button gives info about how to play.
The future version will have three radio buttons for difficulty: easy (+ and - for numbers from 0 to 10), normal (+ and – for numbers from 0 to 20 with timer), and difficult (+ and – for numbers from 0 to 100 with timer).   
It was Tested and approved by my children (9 and 7-years), but maybe only because I made it.